requirements.txt
sqlalchemy==2.0.19
fastapi==0.100.0
faker==19.1.0
uvicorn==0.23.1
pydantic==1.10.10
passlib==1.7.4
libgravatar==1.0.4
alembic==1.11.1
python-dotenv~=1.0.0
cloudinary~=1.33.0
redis~=4.6.0